--- app/models/AddAnApiSelectEndpoints2.scala
@@ -1,44 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package models
-
-import models.api.{ApiDetail, Endpoint, EndpointMethod}
-
-case class AvailableEndpoint(endpoint: Endpoint, endpointMethod: EndpointMethod)
-
-object EndpointMethodScopes {
-
-  def build(apiDetail: ApiDetail): Map[Set[String], AvailableEndpoint] = {
-    apiDetail
-      .endpoints
-      .flatMap(
-        endpoint =>
-          endpoint.methods.map(
-            endpointMethod =>
-              (AvailableEndpoint(endpoint, endpointMethod), endpointMethod.scopes.toSet)
-          )
-      )
-      .map(endpoint => endpoint._2 -> endpoint._1)
-      .toMap
-  }
-
-  // So the page is Set[Set[String]]
-}
-
-object AddAnApiSelectEndpoints2 {
-
-}
