+++ app/pages/AddAnApiSelectEndpointsPage.scala
@@ -19,7 +19,7 @@ package pages
 import models.AddAnApiSelectEndpoints
 import play.api.libs.json.JsPath
 
-case object AddAnApiSelectEndpointsPage extends QuestionPage[Set[AddAnApiSelectEndpoints]] {
+case object AddAnApiSelectEndpointsPage extends QuestionPage[Set[Set[String]]] {
 
   override def path: JsPath = JsPath \ toString
